+++ CHANGELOG.md
@@ -2,5 +2,11 @@
 
 ### Features
 
-- Initial release
-- Basic registry functionality
+- Add enchantment registry.
+- Infer registry entry `Identifier` from field name.
+
+### Fixes
+
+Closed issues: None.
+
+Full Changelog: [Full Changelog](https://github.com/JamCoreModding/JamLib/compare/0.0.1...0.0.2)